+++ samples/docs/generic_platformer_tileset.tsx
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset name="generic_platformer_tileset" tilewidth="32" tileheight="32" tilecount="190" columns="10">
+ <image source="generic_platformer_tileset.png" width="320" height="608"/>
+</tileset>